Update: Adjusted Routes and seperated out logic of route to util file
